This was an attempt to debug the failure of MassEffect/ME2 to load the CLR
There are a lot of changes here, and most are irrelevant; the crux of the matter is that in these games, the CLR fails to start.
Specfically, the call to rHost->Start() fails with E_FAIL.  I can't find any reason for this and I suspect its SecuROM or some other DRM in the game, since every other game I've tried is able to start the CLR with the exact same code.

The call to Start() will fail even if I am not loading my custom app domain.  It also doesn't seem to matter how early I call it; even if I call it right in DllMain's Init() (generally a bad idea), the fail code is returned.  In GW, the game will hang if CLR is initted here, which implies that the ME failure is happening even earlier.

This code includes experimental code to use the core CLR instead of the installed .Net runtime.  This CLR will start at least, however, I can't install my app domain because that portion of the Clr control returns E_NOTIMPL.  Anyway I'm pretty sure F# doesn't work with the core CLR yet and I bet it would fail anyway at some point due to mono or one of its deps not being compatible.  Still, the fact that the Core CLR starts is promising and suggests that eventually we may be able to use it instead of the installed .Net runtime for offensive games like this.

I left this in the state where the CLR is not initialized immediately, but rather when the first reload key is pressed in game.  This is so that I can attach the debugger (steam won't launch the game if a debugger is present at startup).