Allow pixelshader to be specified in mod
This "works", and yet it doesn't work.  I think the issue is that the override pixel shader will _always_ be used when drawing the mesh, but, the game can use multiple shaders for different passes (e.g. shadows), so in fact, you only want to use the mod shader when the base object is being drawn (or in certain reflection passes, etc).  Therefore additional criteria is needed to constrain use of the mod pixel shader.

OY.