Add asynchronous mod loading
Blocking the render thread for X seconds is rude, and could cause bugs.  It gets especially painful as the number of mods increases.

Now the .Net thread pool is used to load.  The render thread will still be blocked for a brief period at the end when the D3D data needs to be created.

Mods are not incrementally _displayed_ as they are loaded.  That requires a deeper change.