More WIP on d3d11 function hooks
- Hook some of the functions of the d3d11 context
- Fix the hook definitions to include the THIS pointer
- Add hook impls for all draw functions that just pass through right now
- Add a hook_release impl that passes through and logs
- Areas I need to revisit marked with TODO11
- The draw functions hooks are currently disabled though (see below)

Remarks:

This "works" although it hangs periodically on startup, however Reshade does the same, so there is likely some implicit race condition with hooking that we both share.

However none of the draw functions get called.  hook_release does get called, exactly once and reports a ref count of 1 at that time.

Possible explanations:

I'm hooking the wrong context? Maybe one created by DXGI for enumeration purposes (see below).  The ref count that is logged from hook_release is 1 which doesn't seem real to me for an actual in-use object (in d3d9 the ref count on the device gets into the thousands or more).

My Vtbl definitions for the d3d11 context are effed up?  IUnknown Release, works but there is a fair amount of functions between that and the context draw functions, including "device child" functions.  I could be stomping some memory badly but in that case I think the game would just crash (and sometimes it starts fine with all funcs hooked).  Could also be a type mismatch in one of the parameters in the winapi D3D11 context rust declarations?

CreateDXGIFactory1 is somehow implicated?  Reshade shows it is used by the game, and I don't hook it (it creates the swap chain though, where Present lives, so I may need it).
Hooking that is possible but annoying because its in a separate DLL (reshade uses a single dll with minhook lib to intercept it at run time)

Draw functions just are not used?  I think DX11 has a command queueing structure instead of just immediate mode like d3d9, so its possible render commands are sent that way (and I'd never observe an actual draw call on this context).  However I did observe at least one call to DrawIndexed in Reshade on the game's main menu.