Made D3D11 draw indexed hook work, sort of.
As I described in the earlier commit f128dda, I appeared to be hooking the context successfully, but the hook for DrawIndexed was not getting called, and I know it should be.

I enumerated various possible causes in that commit message, but in the end it turns out to be the vtable, or more specifically, something is overwritting my hooks.  For only some functions though, IUnknown Release is not overwritten, nor is VSSetConstantBuffers (which by the way is the first function in the context-specific table).

I was able to get hook_draw_indexed to get called by rehooking (if needed) in a new hook function for VSSetConstantBuffers.
This will reset the hook functions every time that function is called.  But something still repeatedly clobbers draw indexed (and maybe others), so I could potentially miss some draw calls "in between".  So not totally happy with this and I might still need to move to an entirely different hook scheme (either using minhook directly or running as a reshade addon).  But its enough for me to move forward.  I cannot fathom what kind of code would overwrite a vtable inside an existing object, other than mine, and then only for some functions, but clearly somebody is doing it.  Possibly window 10 game integration.

Initial analysis shows that the primitive counts for modded stuff does match the dx9 variant, so that is good, but the vertex count is missing in this API so I'm going to have to determine it some other way, probably by inspecting the vertex buffers.